Create abstract base class for commands deleting visits
